# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CreatorAI Studio API
  version: 1.0.0
  description: |
    AI-powered content creation platform API.
    Supports natural language commands, multi-agent workflow orchestration,
    asset management, and multi-platform publishing.
  contact:
    name: CreatorAI Studio
    url: https://creatorai.studio
servers:
  - url: http://localhost:3001/api/v1
    description: Local development
  - url: https://api.creatorai.studio/api/v1
    description: Production
security:
  - BearerAuth: []
...
paths:
...
  /assets/{id}/versions:
    post:
      tags: [Assets]
      summary: Create new asset version
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [changeDescription]
              properties:
                changeDescription: { type: string }
                data: { type: object }
      responses:
        '201':
          description: Version created
    get:
      tags: [Assets]
      summary: Get version history
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Version list
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID Token
      description: Firebase Authentication ID token